tests: Fixed issue where tests stopped in the middle
When running the gunicorn server, the terminal did not start the test scripts
and waited for the process to complete. So I moved the server process to background
and killed it after the tests ran. Should work now in Travis-CI

Will have to look into services like "supervisor" as the current solution is
just a hack.